libc: allow selectively enabling jemalloc's zero-filling
Scudo's nature of being a hardened memory allocator meaned that
improper memory accesses are catched more strictly and vendors had to
fix improper memory accesses from their binaries. This used to be a
problem when switching from jemalloc to Scudo, but now, the inverse has
occured: switching from Scudo to jemalloc is crashing some vendor
binaries.

This is mostly due to the fact that Scudo returns sanitized buffer
while jemalloc doesn't and vendor was lazy enough to not handle
uninitialized buffer properly.

As we do not have the source code to fix such bad binaries, add an
environment interface to selectively enable jemalloc's zero-filling.

To make use of this, add "setenv JE_MALLOC_ZERO_FILLING 1" to the
problematic service's rc file.

Change-Id: Ib3428815a9f5da3787a19f8a18f7609be747659b
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: ReveRTX <rev9rt@gmail.com>